Move autoremove.php to new DatabaseCleanupUtils class
